Firmware release v0.2
Firmware v0.2 uses the remaining GPIO on J4 for signalliting the reset on a power LED. This is normally on. When pressing RESTORE for about 3 seconds, the LED will switch off for a short time, indicating that a normal reset will be performed when RESTORE is released. Holding RESTORE for about 5 seconds will initiate an EXROM reset. While the reset, the LED will be off for a short time, too.